Delete Doxygen comments from spdm_crypt_lib implementations
Make appropriate functions static and move function declarations to the implementation.
