+++ include/library/spdm_crypt_lib.h
@@ -20,429 +20,6 @@
 #define LIBSPDM_MAX_AEAD_TAG_SIZE 16
 
 /**
- * Allocates and initializes one HASH_CTX context for subsequent hash use.
- *
- * @return  Pointer to the HASH_CTX context that has been initialized.
- *          If the allocations fails, libspdm_hash_new_func() returns NULL.
- **/
-typedef void *(*libspdm_hash_new_func)(void);
-
-/**
- * Release the specified HASH_CTX context.
- *
- * @param  hash_context Pointer to the HASH_CTX context to be released.
- **/
-typedef void (*libspdm_hash_free_func)(void *hash_context);
-
-/**
- * Initializes user-supplied memory pointed by hash_context as hash context for
- * subsequent use.
- *
- * @param  base_hash_algo  SPDM base_hash_algo
- * @param  hash_context    Pointer to hash context being initialized.
- *
- * @retval true   Hash context initialization succeeded.
- * @retval false  Hash context initialization failed.
- **/
-typedef bool (*libspdm_hash_init_func)(void *hash_context);
-
-/**
- * Makes a copy of an existing hash context.
- *
- * If hash_ctx is NULL, then return false.
- * If new_hash_ctx is NULL, then return false.
- *
- * @param[in]  hash_ctx      Pointer to hash context being copied.
- * @param[out] new_hash_ctx  Pointer to new hash context.
- *
- * @retval true   Hash context copy succeeded.
- * @retval false  Hash context copy failed.
- *
- **/
-typedef bool (*libspdm_hash_duplicate_func)(const void *hash_ctx, void *new_hash_ctx);
-
-/**
- * Digests the input data and updates hash context.
- *
- * This function performs hash digest on a data buffer of the specified size.
- * It can be called multiple times to compute the digest of long or discontinuous data streams.
- * Hash context should be already correctly initialized by hash_init(), and should not be finalized
- * by hash_final(). Behavior with invalid context is undefined.
- *
- * If hash_context is NULL, then return false.
- *
- * @param[in, out]  hash_context   Pointer to the MD context.
- * @param[in]       data           Pointer to the buffer containing the data to be hashed.
- * @param[in]       data_size      Size of data buffer in bytes.
- *
- * @retval true   hash data digest succeeded.
- * @retval false  hash data digest failed.
- **/
-typedef bool (*libspdm_hash_update_func)(void *hash_context, const void *data, size_t data_size);
-
-/**
- * Completes computation of the hash digest value.
- *
- * This function completes hash computation and retrieves the digest value into
- * the specified memory. After this function has been called, the hash context cannot
- * be used again.
- * hash context should be already correctly initialized by hash_init(), and should not be
- * finalized by hash_final(). Behavior with invalid hash context is undefined.
- *
- * If hash_context is NULL, then return false.
- * If hash_value is NULL, then return false.
- *
- * @param[in, out]  hash_context    Pointer to the hash context.
- * @param[out]      hash_value      Pointer to a buffer that receives the hash digest value.
- *
- * @retval true   hash digest computation succeeded.
- * @retval false  hash digest computation failed.
- **/
-typedef bool (*libspdm_hash_final_func)(void *hash_context, uint8_t *hash_value);
-
-/**
- * Computes the hash of a input data buffer.
- *
- * This function performs the hash of a given data buffer, and return the hash value.
- *
- * @param  data        Pointer to the buffer containing the data to be hashed.
- * @param  data_size   Size of data buffer in bytes.
- * @param  hash_value  Pointer to a buffer that receives the hash value.
- *
- * @retval true   hash computation succeeded.
- * @retval false  hash computation failed.
- **/
-typedef bool (*libspdm_hash_all_func)(const void *data, size_t data_size, uint8_t *hash_value);
-
-/**
- * Allocates and initializes one HMAC context for subsequent hash use.
- *
- * @return  Pointer to the HMAC context that has been initialized.
- *          If the allocations fails, libspdm_hmac_new_func() returns NULL.
- **/
-typedef void *(*libspdm_hmac_new_func)(void);
-
-/**
- * Release the specified HMAC context.
- *
- * @param  hmac_ctx  Pointer to the HMAC context to be released.
- **/
-typedef void (*libspdm_hmac_free_func)(void *hmac_ctx);
-
-/**
- * Set user-supplied key for subsequent use. It must be done before any
- * calling to hmac_update().
- *
- * If hmac_ctx is NULL, then return false.
- *
- * @param[out]  hmac_ctx  Pointer to HMAC context.
- * @param[in]   key       Pointer to the user-supplied key.
- * @param[in]   key_size  Key size in bytes.
- *
- * @retval true   The key is set successfully.
- * @retval false  The key is set unsuccessfully.
- *
- **/
-typedef bool (*libspdm_hmac_set_key_func)(void *hmac_ctx, const uint8_t *key, size_t key_size);
-
-/**
- * Makes a copy of an existing HMAC context.
- *
- * If hmac_ctx is NULL, then return false.
- * If new_hmac_ctx is NULL, then return false.
- *
- * @param[in]  hmac_ctx      Pointer to HMAC context being copied.
- * @param[out] new_hmac_ctx  Pointer to new HMAC context.
- *
- * @retval true   HMAC context copy succeeded.
- * @retval false  HMAC context copy failed.
- *
- **/
-typedef bool (*libspdm_hmac_duplicate_func)(const void *hmac_ctx, void *new_hmac_ctx);
-
-/**
- * Digests the input data and updates HMAC context.
- *
- * This function performs HMAC digest on a data buffer of the specified size.
- * It can be called multiple times to compute the digest of long or discontinuous data streams.
- * HMAC context should be initialized by hmac_new(), and should not be finalized
- * by hmac_final(). Behavior with invalid context is undefined.
- *
- * If hmac_ctx is NULL, then return false.
- *
- * @param[in, out]  hmac_ctx   Pointer to the HMAC context.
- * @param[in]       data       Pointer to the buffer containing the data to be digested.
- * @param[in]       data_size  Size of data buffer in bytes.
- *
- * @retval true   HMAC data digest succeeded.
- * @retval false  HMAC data digest failed.
- *
- **/
-typedef bool (*libspdm_hmac_update_func)(void *hmac_ctx, const void *data, size_t data_size);
-
-/**
- * Completes computation of the HMAC digest value.
- *
- * This function completes HMAC hash computation and retrieves the digest value into
- * the specified memory. After this function has been called, the HMAC context cannot
- * be used again.
- *
- * If hmac_ctx is NULL, then return false.
- * If hmac_value is NULL, then return false.
- *
- * @param[in, out]  hmac_ctx    Pointer to the HMAC context.
- * @param[out]      hmac_value  Pointer to a buffer that receives the HMAC digest value.
- *
- * @retval true   HMAC digest computation succeeded.
- * @retval false  HMAC digest computation failed.
- *
- **/
-typedef bool (*libspdm_hmac_final_func)(void *hmac_ctx, uint8_t *hmac_value);
-
-/**
- * Computes the HMAC of a input data buffer.
- *
- * This function performs the HMAC of a given data buffer, and return the hash value.
- *
- * @param  data        Pointer to the buffer containing the data to be HMACed.
- * @param  data_size   Size of data buffer in bytes.
- * @param  key         Pointer to the user-supplied key.
- * @param  key_size    Key size in bytes.
- * @param  hash_value  Pointer to a buffer that receives the HMAC value.
- *
- * @retval true   HMAC computation succeeded.
- * @retval false  HMAC computation failed.
- **/
-typedef bool (*libspdm_hmac_all_func)(const void *data, size_t data_size,
-                                      const uint8_t *key, size_t key_size,
-                                      uint8_t *hmac_value);
-
-/**
- * Derive HMAC-based Extract key Derivation Function (HKDF) Extract.
- *
- * @param  ikm           Pointer to the input key material.
- * @param  ikm_size      Key size in bytes.
- * @param  salt          Pointer to the salt value.
- * @param  salt_size     Salt size in bytes.
- * @param  prk_out       Pointer to buffer to receive hkdf value.
- * @param  prk_out_size  Size of hkdf bytes to generate.
- *
- * @retval true   Hkdf generated successfully.
- * @retval false  Hkdf generation failed.
- **/
-typedef bool (*libspdm_hkdf_extract_func)(const uint8_t *ikm, size_t ikm_size,
-                                          const uint8_t *salt, size_t salt_size,
-                                          uint8_t *prk_out, size_t prk_out_size);
-
-/**
- * Derive HMAC-based Expand key Derivation Function (HKDF) Expand.
- *
- * @param  prk        Pointer to the user-supplied key.
- * @param  prk_size   Key size in bytes.
- * @param  info       Pointer to the application specific info.
- * @param  info_size  Info size in bytes.
- * @param  out        Pointer to buffer to receive hkdf value.
- * @param  out_size   Size of hkdf bytes to generate.
- *
- * @retval true   Hkdf generated successfully.
- * @retval false  Hkdf generation failed.
- **/
-typedef bool (*libspdm_hkdf_expand_func)(const uint8_t *prk, size_t prk_size,
-                                         const uint8_t *info, size_t info_size,
-                                         uint8_t *out, size_t out_size);
-
-/**
- * Retrieve the asymmetric public key from one DER-encoded X509 certificate.
- *
- * @param  cert       Pointer to the DER-encoded X509 certificate.
- * @param  cert_size  Size of the X509 certificate in bytes.
- * @param  context    Pointer to new-generated asymmetric context which contain the retrieved public
- *                    key component. Use libspdm_asym_free() function to free the resource.
- *
- * @retval  true   public key was retrieved successfully.
- * @retval  false  Fail to retrieve public key from X509 certificate.
- **/
-typedef bool (*libspdm_asym_get_public_key_from_x509_func)(const uint8_t *cert,
-                                                           size_t cert_size,
-                                                           void **context);
-
-/**
- * Release the specified asymmetric context.
- *
- * @param  context  Pointer to the asymmetric context to be released.
- **/
-typedef void (*libspdm_asym_free_func)(void *context);
-
-/**
- * Verifies the asymmetric signature.
- *
- * For RSA/ECDSA, param is NULL.
- * For EdDSA, param is EdDSA context.
- * For EdDSA25519, param is NULL.
- * For EdDSA448, param is EdDSA448 context.
- * For SM2_DSA, param is SM2 IDa.
- *
- * @param  context       Pointer to asymmetric context for signature verification.
- * @param  hash_nid      Hash NID
- * @param  param         Algorithm specific parameter
- * @param  param_size    Algorithm specific parameter size
- * @param  message       Pointer to octet message to be checked (before hash).
- * @param  message_size  Size of the message in bytes.
- * @param  signature     Pointer to asymmetric signature to be verified.
- * @param  sig_size      Size of signature in bytes.
- *
- * @retval  true   Valid asymmetric signature.
- * @retval  false  Invalid asymmetric signature or invalid asymmetric context.
- **/
-typedef bool (*libspdm_asym_verify_func)(void *context, size_t hash_nid,
-                                         const uint8_t *param, size_t param_size,
-                                         const uint8_t *message,
-                                         size_t message_size,
-                                         const uint8_t *signature,
-                                         size_t sig_size);
-
-
-
-/**
- * Carries out the signature generation.
- *
- * If the signature buffer is too small to hold the contents of signature, false
- * is returned and sig_size is set to the required buffer size to obtain the signature.
- *
- * For RSA/ECDSA/EdDSA25519, param is NULL.
- * For EdDSA448, param is EdDSA448 context.
- * For SM2_DSA, param is SM2 IDa.
- *
- * @param  context       Pointer to asymmetric context for signature generation.
- * @param  hash_nid      Hash NID
- * @param  param         Algorithm specific parameter
- * @param  param_size    Algorithm specific parameter size
- * @param  message       Pointer to octet message to be signed (before hash).
- * @param  message_size  Size of the message in bytes.
- * @param  signature     Pointer to buffer to receive signature.
- * @param  sig_size      On input, the size of signature buffer in bytes.
- *                       On output, the size of data returned in signature buffer in bytes.
- *
- * @retval  true   Signature successfully generated.
- * @retval  false  Signature generation failed.
- * @retval  false  Sig_size is too small.
- **/
-typedef bool (*libspdm_asym_sign_func)(void *context, size_t hash_nid,
-                                       const uint8_t *param, size_t param_size,
-                                       const uint8_t *message,
-                                       size_t message_size, uint8_t *signature,
-                                       size_t *sig_size);
-
-/**
- * Allocates and Initializes one Diffie-Hellman Ephemeral (DHE) context for subsequent use.
- *
- * @param nid cipher NID
- *
- * @return  Pointer to the Diffie-Hellman context that has been initialized.
- **/
-typedef void *(*libspdm_dhe_new_by_nid_func)(size_t nid);
-
-/**
- * Generates DHE public key.
- *
- * This function generates random secret exponent, and computes the public key, which is
- * returned via parameter public_key and public_key_size. DH context is updated accordingly.
- * If the public_key buffer is too small to hold the public key, false is returned and
- * public_key_size is set to the required buffer size to obtain the public key.
- *
- * @param  context          Pointer to the DHE context.
- * @param  public_key       Pointer to the buffer to receive generated public key.
- * @param  public_key_size  On input, the size of public_key buffer in bytes.
- *                          On output, the size of data returned in public_key buffer in bytes.
- *
- * @retval true   DHE public key generation succeeded.
- * @retval false  DHE public key generation failed.
- * @retval false  public_key_size is not large enough.
- **/
-typedef bool (*libspdm_dhe_generate_key_func)(void *context,
-                                              uint8_t *public_key,
-                                              size_t *public_key_size);
-
-/**
- * Computes exchanged common key.
- *
- * Given peer's public key, this function computes the exchanged common key, based on its own
- * context including value of prime modulus and random secret exponent.
- *
- * @param  context               Pointer to the DHE context.
- * @param  peer_public_key       Pointer to the peer's public key.
- * @param  peer_public_key_size  Size of peer's public key in bytes.
- * @param  key                   Pointer to the buffer to receive generated key.
- * @param  key_size              On input, the size of key buffer in bytes.
- *                               On output, the size of data returned in key buffer in bytes.
- *
- * @retval true   DHE exchanged key generation succeeded.
- * @retval false  DHE exchanged key generation failed.
- * @retval false  Key_size is not large enough.
- **/
-typedef bool (*libspdm_dhe_compute_key_func)(void *context,
-                                             const uint8_t *peer_public,
-                                             size_t peer_public_size,
-                                             uint8_t *key, size_t *key_size);
-
-/**
- * Release the specified DHE context.
- *
- * @param  context  Pointer to the DHE context to be released.
- **/
-typedef void (*libspdm_dhe_free_func)(void *context);
-
-/**
- * Performs AEAD authenticated encryption on a data buffer and additional authenticated data (AAD).
- *
- * @param  key            Pointer to the encryption key.
- * @param  key_size       Size of the encryption key in bytes.
- * @param  iv             Pointer to the IV value.
- * @param  iv_size        Size of the IV value in bytes.
- * @param  a_data         Pointer to the additional authenticated data (AAD).
- * @param  a_data_size    Size of the additional authenticated data (AAD) in bytes.
- * @param  data_in        Pointer to the input data buffer to be encrypted.
- * @param  data_in_size   Size of the input data buffer in bytes.
- * @param  tag_out        Pointer to a buffer that receives the authentication tag output.
- * @param  tag_size       Size of the authentication tag in bytes.
- * @param  data_out       Pointer to a buffer that receives the encryption output.
- * @param  data_out_size  Size of the output data buffer in bytes.
- *
- * @retval true   AEAD authenticated encryption succeeded.
- * @retval false  AEAD authenticated encryption failed.
- **/
-typedef bool (*libspdm_aead_encrypt_func)(
-    const uint8_t *key, size_t key_size, const uint8_t *iv,
-    size_t iv_size, const uint8_t *a_data, size_t a_data_size,
-    const uint8_t *data_in, size_t data_in_size, uint8_t *tag_out,
-    size_t tag_size, uint8_t *data_out, size_t *data_out_size);
-
-/**
- * Performs AEAD authenticated decryption on a data buffer and additional authenticated data (AAD).
- *
- * @param  key            Pointer to the encryption key.
- * @param  key_size       Size of the encryption key in bytes.
- * @param  iv             Pointer to the IV value.
- * @param  iv_size        Size of the IV value in bytes.
- * @param  a_data         Pointer to the additional authenticated data (AAD).
- * @param  a_data_size    Size of the additional authenticated data (AAD) in bytes.
- * @param  data_in        Pointer to the input data buffer to be decrypted.
- * @param  data_in_size   Size of the input data buffer in bytes.
- * @param  tag            Pointer to a buffer that contains the authentication tag.
- * @param  tag_size       Size of the authentication tag in bytes.
- * @param  data_out       Pointer to a buffer that receives the decryption output.
- * @param  data_out_size  Size of the output data buffer in bytes.
- *
- * @retval true   AEAD authenticated decryption succeeded.
- * @retval false  AEAD authenticated decryption failed.
- **/
-typedef bool (*libspdm_aead_decrypt_func)(
-    const uint8_t *key, size_t key_size, const uint8_t *iv,
-    size_t iv_size, const uint8_t *a_data, size_t a_data_size,
-    const uint8_t *data_in, size_t data_in_size, const uint8_t *tag,
-    size_t tag_size, uint8_t *data_out, size_t *data_out_size);
-
-/**
  * This function returns the SPDM hash algorithm size.
  *
  * @param  base_hash_algo SPDM base_hash_algo
@@ -459,15 +36,6 @@ uint32_t libspdm_get_hash_size(uint32_t base_hash_algo);
  * @return hash cipher ID
  **/
 size_t libspdm_get_hash_nid(uint32_t base_hash_algo);
-
-/**
- * Return asym NID, based upon the negotiated asym algorithm.
- *
- * @param  base_asym_algo  SPDM base_asym_algo
- *
- * @return asym NID
- **/
-size_t libspdm_get_aysm_nid(uint32_t base_asym_algo);
 
 /**
  * Allocates and initializes one HASH_CTX context for subsequent hash use.
@@ -673,23 +241,6 @@ bool libspdm_hash_all(uint32_t base_hash_algo, const void *data,
 uint32_t libspdm_get_measurement_hash_size(uint32_t measurement_hash_algo);
 
 /**
- * Computes the hash of a input data buffer, based upon the negotiated measurement hash algorithm.
- *
- * This function performs the hash of a given data buffer, and return the hash value.
- *
- * @param  measurement_hash_algo  SPDM measurement_hash_algo
- * @param  data                   Pointer to the buffer containing the data to be hashed.
- * @param  data_size              Size of data buffer in bytes.
- * @param  hash_value             Pointer to a buffer that receives the hash value.
- *
- * @retval true   Hash computation succeeded.
- * @retval false  Hash computation failed.
- **/
-bool libspdm_measurement_hash_all(uint32_t measurement_hash_algo,
-                                  const void *data, size_t data_size,
-                                  uint8_t *hash_value);
-
-/**
  * Computes the HMAC of a input data buffer, based upon the negotiated HMAC algorithm.
  *
  * This function performs the HMAC of a given data buffer, and return the hash value.
@@ -823,8 +374,6 @@ bool libspdm_asym_verify_hash(
     size_t hash_size, const uint8_t *signature,
     size_t sig_size);
 
-
-
 /**
  * Carries out the signature generation.
  *
@@ -954,8 +503,6 @@ bool libspdm_req_asym_verify_hash(
     uint32_t base_hash_algo, void *context,
     const uint8_t *message_hash, size_t hash_size,
     const uint8_t *signature, size_t sig_size);
-
-
 
 /**
  * Carries out the signature generation.
@@ -1251,11 +798,10 @@ bool libspdm_get_dmtf_subject_alt_name_from_bytes(
  * @retval true    Get the subjectAltName string successfully
  * @retval failed  Get the subjectAltName string failed
  **/
-bool
-libspdm_get_dmtf_subject_alt_name(const uint8_t *cert, const size_t cert_size,
-                                  char *name_buffer,
-                                  size_t *name_buffer_size,
-                                  uint8_t *oid, size_t *oid_size);
+bool libspdm_get_dmtf_subject_alt_name(const uint8_t *cert, const size_t cert_size,
+                                       char *name_buffer,
+                                       size_t *name_buffer_size,
+                                       uint8_t *oid, size_t *oid_size);
 
 /**
  * This function verifies the integrity of certificate chain data without spdm_cert_chain_t header.
