+++ library/spdm_crypt_lib/libspdm_crypt_hash.c
@@ -7,12 +7,100 @@
 #include "internal/libspdm_crypt_lib.h"
 
 /**
- * This function returns the SPDM hash algorithm size.
+ * Allocates and initializes one HASH_CTX context for subsequent hash use.
  *
- * @param  base_hash_algo                  SPDM base_hash_algo
- *
- * @return SPDM hash algorithm size.
+ * @return  Pointer to the HASH_CTX context that has been initialized.
+ *          If the allocations fails, libspdm_hash_new_func() returns NULL.
  **/
+typedef void *(*libspdm_hash_new_func)(void);
+
+/**
+ * Release the specified HASH_CTX context.
+ *
+ * @param  hash_context Pointer to the HASH_CTX context to be released.
+ **/
+typedef void (*libspdm_hash_free_func)(void *hash_context);
+
+/**
+ * Initializes user-supplied memory pointed by hash_context as hash context for
+ * subsequent use.
+ *
+ * @param  base_hash_algo  SPDM base_hash_algo
+ * @param  hash_context    Pointer to hash context being initialized.
+ *
+ * @retval true   Hash context initialization succeeded.
+ * @retval false  Hash context initialization failed.
+ **/
+typedef bool (*libspdm_hash_init_func)(void *hash_context);
+
+/**
+ * Makes a copy of an existing hash context.
+ *
+ * If hash_ctx is NULL, then return false.
+ * If new_hash_ctx is NULL, then return false.
+ *
+ * @param[in]  hash_ctx      Pointer to hash context being copied.
+ * @param[out] new_hash_ctx  Pointer to new hash context.
+ *
+ * @retval true   Hash context copy succeeded.
+ * @retval false  Hash context copy failed.
+ *
+ **/
+typedef bool (*libspdm_hash_duplicate_func)(const void *hash_ctx, void *new_hash_ctx);
+
+/**
+ * Digests the input data and updates hash context.
+ *
+ * This function performs hash digest on a data buffer of the specified size.
+ * It can be called multiple times to compute the digest of long or discontinuous data streams.
+ * Hash context should be already correctly initialized by hash_init(), and should not be finalized
+ * by hash_final(). Behavior with invalid context is undefined.
+ *
+ * If hash_context is NULL, then return false.
+ *
+ * @param[in, out]  hash_context   Pointer to the MD context.
+ * @param[in]       data           Pointer to the buffer containing the data to be hashed.
+ * @param[in]       data_size      Size of data buffer in bytes.
+ *
+ * @retval true   hash data digest succeeded.
+ * @retval false  hash data digest failed.
+ **/
+typedef bool (*libspdm_hash_update_func)(void *hash_context, const void *data, size_t data_size);
+
+/**
+ * Completes computation of the hash digest value.
+ *
+ * This function completes hash computation and retrieves the digest value into
+ * the specified memory. After this function has been called, the hash context cannot
+ * be used again.
+ * hash context should be already correctly initialized by hash_init(), and should not be
+ * finalized by hash_final(). Behavior with invalid hash context is undefined.
+ *
+ * If hash_context is NULL, then return false.
+ * If hash_value is NULL, then return false.
+ *
+ * @param[in, out]  hash_context    Pointer to the hash context.
+ * @param[out]      hash_value      Pointer to a buffer that receives the hash digest value.
+ *
+ * @retval true   hash digest computation succeeded.
+ * @retval false  hash digest computation failed.
+ **/
+typedef bool (*libspdm_hash_final_func)(void *hash_context, uint8_t *hash_value);
+
+/**
+ * Computes the hash of a input data buffer.
+ *
+ * This function performs the hash of a given data buffer, and return the hash value.
+ *
+ * @param  data        Pointer to the buffer containing the data to be hashed.
+ * @param  data_size   Size of data buffer in bytes.
+ * @param  hash_value  Pointer to a buffer that receives the hash value.
+ *
+ * @retval true   hash computation succeeded.
+ * @retval false  hash computation failed.
+ **/
+typedef bool (*libspdm_hash_all_func)(const void *data, size_t data_size, uint8_t *hash_value);
+
 uint32_t libspdm_get_hash_size(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
@@ -32,13 +120,6 @@ uint32_t libspdm_get_hash_size(uint32_t base_hash_algo)
     }
 }
 
-/**
- * Return cipher ID, based upon the negotiated hash algorithm.
- *
- * @param  base_hash_algo                  SPDM base_hash_algo
- *
- * @return hash cipher ID
- **/
 size_t libspdm_get_hash_nid(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
@@ -68,7 +149,7 @@ size_t libspdm_get_hash_nid(uint32_t base_hash_algo)
  *
  * @return hash new function
  **/
-libspdm_hash_new_func libspdm_get_hash_new_func(uint32_t base_hash_algo)
+static libspdm_hash_new_func libspdm_get_hash_new_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -135,7 +216,7 @@ libspdm_hash_new_func libspdm_get_hash_new_func(uint32_t base_hash_algo)
  *
  * @return hash free function
  **/
-libspdm_hash_free_func libspdm_get_hash_free_func(uint32_t base_hash_algo)
+static libspdm_hash_free_func libspdm_get_hash_free_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -202,7 +283,7 @@ libspdm_hash_free_func libspdm_get_hash_free_func(uint32_t base_hash_algo)
  *
  * @return hash init function
  **/
-libspdm_hash_init_func libspdm_get_hash_init_func(uint32_t base_hash_algo)
+static libspdm_hash_init_func libspdm_get_hash_init_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -262,7 +343,6 @@ libspdm_hash_init_func libspdm_get_hash_init_func(uint32_t base_hash_algo)
     return NULL;
 }
 
-
 /**
  * Return hash duplicate function, based upon the negotiated hash algorithm.
  *
@@ -270,7 +350,7 @@ libspdm_hash_init_func libspdm_get_hash_init_func(uint32_t base_hash_algo)
  *
  * @return hash duplicate function
  **/
-libspdm_hash_duplicate_func libspdm_get_hash_duplicate_func(uint32_t base_hash_algo)
+static libspdm_hash_duplicate_func libspdm_get_hash_duplicate_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -337,7 +417,7 @@ libspdm_hash_duplicate_func libspdm_get_hash_duplicate_func(uint32_t base_hash_a
  *
  * @return hash update function
  **/
-libspdm_hash_update_func libspdm_get_hash_update_func(uint32_t base_hash_algo)
+static libspdm_hash_update_func libspdm_get_hash_update_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -404,7 +484,7 @@ libspdm_hash_update_func libspdm_get_hash_update_func(uint32_t base_hash_algo)
  *
  * @return hash final function
  **/
-libspdm_hash_final_func libspdm_get_hash_final_func(uint32_t base_hash_algo)
+static libspdm_hash_final_func libspdm_get_hash_final_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -471,7 +551,7 @@ libspdm_hash_final_func libspdm_get_hash_final_func(uint32_t base_hash_algo)
  *
  * @return hash function
  **/
-libspdm_hash_all_func libspdm_get_hash_all_func(uint32_t base_hash_algo)
+static libspdm_hash_all_func libspdm_get_hash_all_func(uint32_t base_hash_algo)
 {
     switch (base_hash_algo) {
     case SPDM_ALGORITHMS_BASE_HASH_ALGO_TPM_ALG_SHA_256:
@@ -531,14 +611,6 @@ libspdm_hash_all_func libspdm_get_hash_all_func(uint32_t base_hash_algo)
     return NULL;
 }
 
-/**
- * Allocates and initializes one HASH_CTX context for subsequent hash use.
- *
- * @param  base_hash_algo                 SPDM base_hash_algo
- *
- * @return  Pointer to the HASH_CTX context that has been initialized.
- *         If the allocations fails, libspdm_hash_new() returns NULL.
- **/
 void *libspdm_hash_new(uint32_t base_hash_algo)
 {
     libspdm_hash_new_func hash_function;
@@ -549,12 +621,6 @@ void *libspdm_hash_new(uint32_t base_hash_algo)
     return hash_function();
 }
 
-/**
- * Release the specified HASH_CTX context.
- *
- * @param  base_hash_algo                 SPDM base_hash_algo
- * @param  hash_context                   Pointer to the HASH_CTX context to be released.
- **/
 void libspdm_hash_free(uint32_t base_hash_algo, void *hash_context)
 {
     libspdm_hash_free_func hash_function;
@@ -565,16 +631,6 @@ void libspdm_hash_free(uint32_t base_hash_algo, void *hash_context)
     hash_function(hash_context);
 }
 
-/**
- * Initializes user-supplied memory pointed by hash_context as hash context for
- * subsequent use.
- *
- * @param  base_hash_algo                 SPDM base_hash_algo
- * @param  hash_context                   Pointer to hash context being initialized.
- *
- * @retval true   Hash context initialization succeeded.
- * @retval false  Hash context initialization failed.
- **/
 bool libspdm_hash_init(uint32_t base_hash_algo, void *hash_context)
 {
     libspdm_hash_init_func hash_function;
@@ -585,19 +641,6 @@ bool libspdm_hash_init(uint32_t base_hash_algo, void *hash_context)
     return hash_function(hash_context);
 }
 
-/**
- * Makes a copy of an existing hash context.
- *
- * If hash_ctx is NULL, then return false.
- * If new_hash_ctx is NULL, then return false.
- *
- * @param[in]  hash_ctx     Pointer to hash context being copied.
- * @param[out] new_hash_ctx  Pointer to new hash context.
- *
- * @retval true   hash context copy succeeded.
- * @retval false  hash context copy failed.
- *
- **/
 bool libspdm_hash_duplicate(uint32_t base_hash_algo, const void *hash_ctx, void *new_hash_ctx)
 {
     libspdm_hash_duplicate_func hash_function;
@@ -608,23 +651,6 @@ bool libspdm_hash_duplicate(uint32_t base_hash_algo, const void *hash_ctx, void 
     return hash_function(hash_ctx, new_hash_ctx);
 }
 
-/**
- * Digests the input data and updates hash context.
- *
- * This function performs hash digest on a data buffer of the specified size.
- * It can be called multiple times to compute the digest of long or discontinuous data streams.
- * Hash context should be already correctly initialized by hash_init(), and should not be finalized
- * by hash_final(). Behavior with invalid context is undefined.
- *
- * If hash_context is NULL, then return false.
- *
- * @param[in, out]  hash_context   Pointer to the MD context.
- * @param[in]       data           Pointer to the buffer containing the data to be hashed.
- * @param[in]       data_size      Size of data buffer in bytes.
- *
- * @retval true   hash data digest succeeded.
- * @retval false  hash data digest failed.
- **/
 bool libspdm_hash_update(uint32_t base_hash_algo, void *hash_context,
                          const void *data, size_t data_size)
 {
@@ -636,24 +662,6 @@ bool libspdm_hash_update(uint32_t base_hash_algo, void *hash_context,
     return hash_function(hash_context, data, data_size);
 }
 
-/**
- * Completes computation of the hash digest value.
- *
- * This function completes hash computation and retrieves the digest value into
- * the specified memory. After this function has been called, the hash context cannot
- * be used again.
- * hash context should be already correctly initialized by hash_init(), and should not be
- * finalized by hash_final(). Behavior with invalid hash context is undefined.
- *
- * If hash_context is NULL, then return false.
- * If hash_value is NULL, then return false.
- *
- * @param[in, out]  hash_context    Pointer to the hash context.
- * @param[out]      hash_value      Pointer to a buffer that receives the hash digest value.
- *
- * @retval true   hash digest computation succeeded.
- * @retval false  hash digest computation failed.
- **/
 bool libspdm_hash_final(uint32_t base_hash_algo, void *hash_context, uint8_t *hash_value)
 {
     libspdm_hash_final_func hash_function;
@@ -664,19 +672,6 @@ bool libspdm_hash_final(uint32_t base_hash_algo, void *hash_context, uint8_t *ha
     return hash_function(hash_context, hash_value);
 }
 
-/**
- * Computes the hash of a input data buffer, based upon the negotiated hash algorithm.
- *
- * This function performs the hash of a given data buffer, and return the hash value.
- *
- * @param  base_hash_algo                 SPDM base_hash_algo
- * @param  data                         Pointer to the buffer containing the data to be hashed.
- * @param  data_size                     size of data buffer in bytes.
- * @param  hash_value                    Pointer to a buffer that receives the hash value.
- *
- * @retval true   hash computation succeeded.
- * @retval false  hash computation failed.
- **/
 bool libspdm_hash_all(uint32_t base_hash_algo, const void *data,
                       size_t data_size, uint8_t *hash_value)
 {
@@ -688,14 +683,6 @@ bool libspdm_hash_all(uint32_t base_hash_algo, const void *data,
     return hash_function(data, data_size, hash_value);
 }
 
-/**
- * This function returns the SPDM measurement hash algorithm size.
- *
- * @param  measurement_hash_algo          SPDM measurement_hash_algo
- *
- * @return SPDM measurement hash algorithm size.
- * @return 0xFFFFFFFF for RAW_BIT_STREAM_ONLY.
- **/
 uint32_t libspdm_get_measurement_hash_size(uint32_t measurement_hash_algo)
 {
     switch (measurement_hash_algo) {
@@ -715,96 +702,4 @@ uint32_t libspdm_get_measurement_hash_size(uint32_t measurement_hash_algo)
     default:
         return 0;
     }
-}
-
-/**
- * Return hash function, based upon the negotiated measurement hash algorithm.
- *
- * @param  measurement_hash_algo          SPDM measurement_hash_algo
- *
- * @return hash function
- **/
-libspdm_hash_all_func libspdm_spdm_measurement_hash_func(uint32_t measurement_hash_algo)
-{
-    switch (measurement_hash_algo) {
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA_256:
-#if LIBSPDM_SHA256_SUPPORT
-        return libspdm_sha256_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA_384:
-#if LIBSPDM_SHA384_SUPPORT
-        return libspdm_sha384_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA_512:
-#if LIBSPDM_SHA512_SUPPORT
-        return libspdm_sha512_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA3_256:
-#if LIBSPDM_SHA3_256_SUPPORT
-        return libspdm_sha3_256_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA3_384:
-#if LIBSPDM_SHA3_384_SUPPORT
-        return libspdm_sha3_384_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SHA3_512:
-#if LIBSPDM_SHA3_512_SUPPORT
-        return libspdm_sha3_512_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    case SPDM_ALGORITHMS_MEASUREMENT_HASH_ALGO_TPM_ALG_SM3_256:
-#if LIBSPDM_SM3_256_SUPPORT
-        return libspdm_sm3_256_hash_all;
-#else
-        LIBSPDM_ASSERT(false);
-        break;
-#endif
-    default:
-        LIBSPDM_ASSERT(false);
-        break;
-    }
-
-    return NULL;
-}
-
-/**
- * Computes the hash of a input data buffer, based upon the negotiated measurement hash algorithm.
- *
- * This function performs the hash of a given data buffer, and return the hash value.
- *
- * @param  measurement_hash_algo          SPDM measurement_hash_algo
- * @param  data                         Pointer to the buffer containing the data to be hashed.
- * @param  data_size                     size of data buffer in bytes.
- * @param  hash_value                    Pointer to a buffer that receives the hash value.
- *
- * @retval true   hash computation succeeded.
- * @retval false  hash computation failed.
- **/
-bool libspdm_measurement_hash_all(uint32_t measurement_hash_algo,
-                                  const void *data, size_t data_size,
-                                  uint8_t *hash_value)
-{
-    libspdm_hash_all_func hash_function;
-    hash_function = libspdm_spdm_measurement_hash_func(measurement_hash_algo);
-    if (hash_function == NULL) {
-        return false;
-    }
-    return hash_function(data, data_size, hash_value);
 }
